Update create-an-odata-v4-client-app.md
Hi Team, I found a bit unnatural sentence in Japanese language, and I would like to fix it. This proposal is supposed to be addressed as a Linguistic Suggestion from my point of view. It would be appreciated if you would check this proposal. Please let me know if you have any concerns.

This request is in regard to:

- [OData v4 クライアント アプリでは (c#) を作成します。](https://docs.microsoft.com/ja-jp/aspnet/web-api/overview/odata-support-in-aspnet-web-api/odata-v4/create-an-odata-v4-client-app)
- [Create an OData v4 Client App (C#)](https://docs.microsoft.com/en-us/aspnet/web-api/overview/odata-support-in-aspnet-web-api/odata-v4/create-an-odata-v4-client-app)

Best regards,